update difference after pr #10
